Read repository URLs from the source pom
If a pom is used as the input file read repository URLs from the
repositories section. If there is a baseUrl value provided on
command-line use it instead to allow override of the values in the pom.
If pom does not contain any repositories nor a baseUrl value is provided
then use the default value.

When using downlog file as an input there is no way how to provide more
than one repository URL ATM.